Drop "gentle" collection mode
Summary:
We no longer need two modes. Now that collection only occurs between rechecks,
we only use the "aggressive" mode, which tries to keep the amount of garbage in
the heap to less than 20%.

The incremental GC will also strive to keep the amount of garbage less than 20%.
Removing the gentle mode now makes it clear that the incremental GC is behavior
preserving.

It might make sense to make the max overhead configurable, but I'm leaving that
for future work.

Reviewed By: nmote

Differential Revision: D25340133

fbshipit-source-id: b780038dac552be9d10356854e61f2e9831de952