Move and rename SharedMem_js module
Summary:
In a previous diff, I removed the wrapping behavior of this module. After this
change, the module only exports a couple of useful memory logging utilities.

This diff simply moves the file to a new Memory_utils module in the inference
target, where the functions are most commonly used.

Faceook:
I suppose feel free to bikeshed on this diff.

Reviewed By: nmote

Differential Revision: D25280249

fbshipit-source-id: de1b9502bbf8a5cd0d6c5361070c2373e6263b09