add def_loc to Ty representation of object members
Summary: `Ty_normalizer.expand_members` is intended to replace `members.ml`, because `members.ml` is deprecated and continues to be super buggy. In order to be able to make get-def use `Ty_normalizer.expand_members`, we need the `Ty.t` representation of object members to carry the location of the member's definition. This diff adds that to the `Ty.t` representation, but doesn't make anything consume it yet. Future diffs will make get-def use this instead of `members.ml`.

Reviewed By: panagosg7

Differential Revision: D25379982

fbshipit-source-id: 6f7cb43781bce8ffd8ff40328e071ff58e636f79