Flow Enums runtime package
Summary:
This diffs moves the existing Flow Enums runtime, as defined and used in WWW, to its own package. We copy over the code and tests.

In the future, when we open-source Flow Enums (announce, and move the docs to flow.org), I'll update and expand the README with more information.

Reviewed By: Hans-Halverson

Differential Revision: D25475422

fbshipit-source-id: 83235bab86a2666bd39633fedd59e93bdcc0b180