Flow Enums transform package
Summary:
This diffs moves the existing Flow Enums Babel transform, as defined and used in WWW, to its own package. We copy over the code and tests.

In the future, when we open-source Flow Enums (announce, and move the docs to flow.org), I'll update and expand the README with more information.

Reviewed By: pieterv

Differential Revision: D25475421

fbshipit-source-id: 615abc36c8106dd967a5176c38c1c6e7900df8a7