merge imports before check
Summary:
Before, when checking a component, we would first generate constraints
for each AST in that component, and then connect each import with the type
of the module that the dependency exports. Moving forward, we would like
to have concrete information about the imports available when visiting the
file's AST during checking. This means that this ordering needs to be reversed.

This change is only meaningful for components of size 1, ie. non-cyclical
components. Luckily, this is the case for the Check phase of types-first. This
diff applies this change in that scenario only. I introduced `check_component`
to do this.

Reviewed By: samwgoldman

Differential Revision: D24112595

fbshipit-source-id: 4a03d9f7548a29b7542eb607029a5769909ca3de