Track changed files in merge stream
Summary: Previously, we computed the set of files that have new or changed signatures by simply taking the set of unrevived files, i.e. the set of files which were considered for merge (or deleted, or not parsed) except those that merge actually ended up modifying. This seems reasonable, but in upcoming diffs I am going to need to change which signatures we consider to be "new" for the purposes of recheck opts, meaning that this needs to get tracked explicitly in merge stream.

Reviewed By: mroch

Differential Revision: D24935279

fbshipit-source-id: 6b4c814cdf5e919a79ea79325f5a9b14ef732f2b