runtests.sh: Do not generate saved state with a lazy server
Summary:
The changes in D25137486 require the saved state to be generated from a full init (or from a saved state init that loaded a saved state generated from a full init, etc.), because the optimizations in that diff require sighashes to be available in the saved state.

Note generating a saved state from a lazy server would not affect correctness, it would only mean that the optimizations added in D25137486 would not be able to fire.

It may make sense to create some sort of mode that *only* merges, so that we can more quickly generate saved state. The check phase doesn't produce any output that's necessary for saved state.

Reviewed By: mroch

Differential Revision: D25137487

fbshipit-source-id: f69fec064e71952e3eca67c8c0331045a1c0fc8f