fix flow error in packages/flow-dev-tools
Summary:
flow-oss-mac and flow-oss-linux were failing, due to errors in flow-dev-tools.
`flow check` errored because not all tests in `newtests/lsp/completion/`
were excluded. Files in this directory are not good candidates when `flow check`
is expected to pass, because some of the files there have syntax errors. To avoid
accidental errors, I'm excluding the entire folder.

Reviewed By: jbrown215

Differential Revision: D25634573

fbshipit-source-id: 18ddf4b7957f34685c195ea00cc933b9b7dba3f5