fix flow error in packages/flow-dev-tools
Summary:
flow-oss-mac and flow-oss-linux were failing, due to errors in flow-dev-tools.
`flow check` errored because not all tests in `newtests/lsp/completion/`
were excluded. Files in this directory are not good candidates when `flow check`
is expected to pass, because some of the files there have syntax errors. To avoid
accidental errors, I'm excluding the entire folder.

Reviewed By: jbrown215

Differential Revision: D25634573 (https://github.com/facebook/flow/commit/033a851943ba23c253acf431ef7ff164a217977b)

fbshipit-source-id: 079246a7ca7b1bb841a52f8530ee3f74a15c48eb