Add Subtyping_kit
Summary:
This diff moves the handling of subtyping constraints (pairs of the form
`(t, UseT t')` out of Flow_js and into a new module Subtyping_kit. This
module is recursively dependent on Flow_js.

Motivation for this change is twofold. On the one hand, it simplifies the
large case analysis in Flow_js, and simplifies the ordering of the remaining
cases. On the other hand, it is a first step towards making subtyping
a self contained function: calls to subtyping should only lead to constraints
within the Subtyping_kit module, instead of other "use_t" constraints in
Flow_js. This doesn't hold at the moment, but by inspecting the various
cases of subtyping we get an idea of the work that needs to be done.

In terms of implementation, this is largely a refactoring that moves cases from
Flow_js to Subtyping_kit, while adding the necessary dependencies to
Flow_common. Some of these operations are still quite high level. In future
work, these will be included in other "kit" modules, so that only primitive
operations (e.g. rec_flow, unify) need to be passed into Subtyping_kit.

Some of the cases that were part of the original Flow_js switch statement
had to be duplicated in Subtyping_kit. These were cases where the RHS of
the constraint was a catch-all (P151711794). In almost all of those cases,
the code on the right side of the "->" was simple enough to justify duplication.
The only exception was the case of UnionT for which I added a `iter_resolve_union`
helper in D25236155.

Reviewed By: jbrown215

Differential Revision: D25171327

fbshipit-source-id: 17ed9a24be6c7ef6eb942ba7e9503020ab22c3b7