[refactor] Move non-recursive functions out of Flow_js
Summary:
Flow_js has grown considerably both in size and in complexity. There is a good
reason for this: this module implements a recursively defined subtyping/propagation
relation. Functions that are part of this recursion and may generate new constraints
need to be defined in this file. All other functions do not need to be here. This diff
moves such functions to flow_js_utils.ml. This change:
* helps remove clutter from flow_js_utils.ml
* allows this function to be shared with other modules added in this stack
* separates *pure* functions from those that may be generating constraints (through
calls to `flow`, etc.).

Reviewed By: jbrown215

Differential Revision: D25171326

fbshipit-source-id: 5ca8ce6f101c00a0911e8a01f74d4ddedbd17e2c