[will cause errors] Do not allow Inexact to Indexed object flows
Summary:
Inexact objects may have properties not enumerated in their types with unknown types for their values. When we flow an inexact object to an indexed object, we should conservatively assume some unknown property exists with type `mixed` to be sound. This effectively bans most inexact ~> indexed object flows.

We did not enforce this in the past because indexers were already unsound and we left the user to enforce certain properties about indexed objects themselves. We are now aware of use-cases that requires us to take a stand with this extra strictness.

Reviewed By: panagosg7

Differential Revision: D24338231

fbshipit-source-id: 2659036ecb88f901ec619c1a30d4a00cf1b7efe0