[will cause errors] Disallow Function ~> Indexed object flows
Summary: Disallows function ~> Indexed object flows. This subtyping relations would almost never have passed anyway, since it requires the value type on the indexer to allow every property on the function's statics _AND_ Function.prototype. This makes it easier for union speculation to make progress when subtyping a function against unions that include objects with indexers. In some cases, like the one in the included tests, it is necessary in order to make progress soundly.

Reviewed By: panagosg7

Differential Revision: D24515951

fbshipit-source-id: b86e2582b9251f1285c4458d0b68f990c5795e1f