[easy] move annotate-exports tests to types-first
Summary:
"annotate-exports" tests refer to the process of upgrading from classic mode
to types-first. Even though we'll soon drop support for classic mode, we might
still want to keep the annotate-exports command around for a bit longer. This
code can be used to fill in output types on a per-file basis, within a types-first repo.

This diff removes the "types_first=false" option in tests that could be directly ported.

Reviewed By: jbrown215

Differential Revision: D25718096

fbshipit-source-id: 029f5f68c2068f838f1adf612e708217b94ace32