update computed_props test for types-first
Summary:
Computed props in exports are not supported in types-first 1.0 (T64194787).
They are supported a little bit better in types-first 2.0, but the behavior between
the two versions diverges. Instead of removing this test completely, this diff
prepares this test for types-first, enables it for new_signatures only, and adds
some TODOs.

Reviewed By: Hans-Halverson

Differential Revision: D25711292

fbshipit-source-id: 2c3d960401e0dbe557a5ba52e517c7c7f615454d