fix get-def tests for types-first
Summary:
Prepare tests for end of support of classic mode.

In types-first, the signature for
```
export let letBinding: string = "";
```
is
```
declare var letBinding: string;
export {letBinding};
```
Before, the location of the second occurrence of `letBinding` in the output was the
entire declaration. With this change this becomes the span of just the identifier. This
matches the expected behavior for get-def in classic mode.

Reviewed By: mroch, vrama628

Differential Revision: D25573671

fbshipit-source-id: 2ecda91d22aafe836734da7b2c202ce99d79cf1b