compute exports during parsing
Summary:
for autoimports, we need a map from every possible name you could import, to the file that exports it.

first, for every file, we need to compute a list of its exports. later, we will invert this file -> exports map to an export -> files map.

we use `Exports.of_type_sig` to get the list of exports from the type sig. I compute this in `Parsing_service_js` where we already compute the type sig.

- in types-first 2.0, we already have the type sig, so this is natural and cheap.
- in types-first 1.0, I am *additionally* computing the type sig, which is a perf hit, but a stopgap until 2.0 is ready.

### Shared mem:

I'm currently storing all the exports in shared mem (`ExportsHeap`), but that's just to have a unified codepath for TF 1.0 and 2.0. once 2.0 is shipped, we can compute the exports on demand from the type sigs in shared mem.

### Saved state:

I'm also persisting `ExportsHeap` in saved state, since we don't store the type sigs. on init, to build the (upcoming) export -> files map, we will walk all of the entries of `ExportsHeap` and build the inverse map, rather than directly storing it; this simplifies incremental updates, because `ExportsHeap` is stored per file, and appears to be much smaller because there are many fewer total files than there are total exports (obviously).

Reviewed By: nmote

Differential Revision: D25418706

fbshipit-source-id: a30bbc5c6f0dabcfe4a48adcc73ed7fbc86aa735