ocaml bindings for fuzzy-native
Summary:
fuzzy-native is a multithreaded C++ library for fuzzy matching used by nuclide: https://github.com/facebook-atom/nuclide-prebuilt-libs/tree/master/fuzzy-native

this vendors in commit 127dfa0bfdaa3403919699ecfaeb93daea84b07b in src/third-party/fuzzy-path/vendor/, and adds ocaml bindings in src/third-party/fuzzy-path/src/.

### planned usage

this will be used for searching the index of exported names to power auto-imports during autocomplete and code actions.

it'll also be used for autocompleting module names, which are paths in node, so the tuning for paths will be useful.

### why C++

there's nothing particularly useful about the core algorithm being implemented in C++, but parallelism makes a significant difference in my tests.

porting to C would make the ocaml integration easier (see below) but the second benefit is that this is used by nuclide and many IDE plugins at Facebook so we should reuse the existing C++.

### ocaml bindings

i called the ocaml library `fuzzy-path`, not `fuzzy-native`, because it's tuned for file paths and the fact that it calls `fuzzy-native` is an implementation detail.

as for implementation: since it is in C++ while `ocamlopt` only knows how to compile C, I created an `extern "C"` API in fuzzy_path_wrapper.cpp, and ocaml stubs in C in fuzzy_path_stubs.c. we have to manually compile using `g++` outside of `ocamlbuild` and have it link it as a prebuilt artifact.

(note: dune knows how to build C++ so this will get better once we finally switch. or it could be split into a standalone repo/opam module using dune)

Reviewed By: nmote

Differential Revision: D25220168

fbshipit-source-id: d622fdfcedf621a07ef3a89121f6f5a98f9c2b46