MethodTs use the left-hand side of their flows as the `this` argument when transforming into CallTs
Summary:
Because of the lax type checking around `this` parameters that exists as a result of their current default of `any`, much of the checking we do around method calls and `this` arguments is somewhat imprecise, and can create spurious incompatibilities between the object on which we look up the method property and the `this` argument we provide to the resulting function call.

Today, when we see a method call of the form `x.m()`, for some `x` of type `X`, we generate a flow of the form
 `X ~> MethodT ("m", this_arg = X)`. This decomposes into
`X -> GetPropT ("m", tout)` and `tout -> CallT (this_arg = X)` flows. Most of the time this is correct, but not always. Consider what happens when `X` is a union of objects, and the type of `m` depends on which branch of the union we are in when looking up the method type. The following construction illustrates this behavior:

```
function foo () {
    return this.x;
}
let o1 = { foo, x : 3 }

function bar() {
    return this.y;
}
let o2 = { foo : bar, y : 3}
declare var x : typeof o1 | typeof o2;
```

In this case, `X` is a union of two types, both of which define a method called `foo`, but that have type signatures that differ in the `this` parameter. When `o1 | o2 ~> MethodT ("foo")` decomposes into
`o1 | o2 ~> GetPropT ("foo", tout)` and `tout -> CallT(this_arg = o1 | o2)`, the `GetPropT` flow will split into two, giving `tout` two lower bounds, one of type `(this : {x : number}) => number` and the other of type `(this : {y : number}) => number`. Both of these will then flow to the `CallT` which will flow the `this_arg` of type `o1 | o2` to both of these functions, and raise errors for both cases.

Instead we would like to use the same `this` parameter for the function call as we used to look up the method, since the latter is able to influence the former in cases like this.

This diff also generalizes the logic in the previous diff to promote the `this` argument to methods on primitives.

Reviewed By: mvitousek

Differential Revision: D24894354

fbshipit-source-id: 23ab00fc760c1a665d12af297ae9057465a78198