add index of all files' exports
Summary:
tl;dr: processes the file -> exports index stored in shared mem into a export -> file index stored in master

--

`ExportsHeap` stores the index from file -> exported names in shared memory. in order to search by exported name for auto-imports, we'd have to map/reduce over ExportsHeap using the workers, but autoimports need to run in autocomplete and code actions, which are parallelizable with rechecks. so we can't use workers, and the overhead of spawning workers on every keystroke is too high.

instead, we invert the map into exported name -> files and store it in the main process (specifically `ServerEnv`), and use `fuzzy-path` to search it in parallel with threads instead of workers.

### `Export_index`

Stores the exported name -> (file * kind of import) map

### `Export_search`

This coordinates between `Export_index` and `fuzzy-path`:

- `fuzzy-path` only knows about a set of strings, so this looks those results up in the index to also return filenames for each result.
- as the index is updated, adds/removes entries from `fuzzy-path`

### `Export_service`

As with other "services", this manages the workers. Given a list of files, it has the workers read from `ExportsHeap` and return an `Export_index`, which all get merged in the main process.

Reviewed By: nmote, vrama628

Differential Revision: D25418704

fbshipit-source-id: 38f7183abcf99ba500c66c98f02d08f25c836a7d