distinguish between methods and function properties in objects
Summary: In order to diverge the typechecking behavior of methods and function properties with respect to their `this` parameters, we must ensure that they are properly different in Flow. Previously, adding a method to an object would simply generate a positive polarity property with a function type. Now we correctly store this as a method.

Reviewed By: mvitousek

Differential Revision: D25592386

fbshipit-source-id: 95c77fee53994bec9873efbe70a01b2ce89674f2