type checking for this annotations on class methods
Summary:
This implements typechecking for `this` parameters in class methods, declared class methods and interfaces. This implements no inference, and does not change the behavior of any code that does not use `this` annotations.

There are two important behaviors to note for `this` parameters of class methods:

* The `this` parameter of a class method must be a supertype of the instance/static (depending on whether the method is static or not) type of the class. More concretely: for any class of the following form:
     class X {
         foo (this : A) { ... }
         static bar (this : B) { ... }
     }
the following conditions must hold: `X <: A` and `Class<X> <: B`. These two are necessary for the second behavior:

   * `this` parameters are of overridden methods are not compared during subclassing checks. Dynamic dispatch, along with the above restrictions, ensures that without resorting to unbinding and rebinding class methods, we cannot call a class method in such a way as to be unsound w.r.t to the `this` parameter. This obviates the need to compare the `this` parameter when overriding a class method.

Reviewed By: mvitousek

Differential Revision: D24334235

fbshipit-source-id: 91563ef61fc95d3a8e6c483353696bca32cb0ea3