store this type param separately from other class tparams
Summary:
Classes treat `this` as a special type parameter and add it to their `tparams_map` and `tparams` list on creation, whereas interfaces have no concept of `this` and do not include it in their map. Currently this is implemented simply by adding the `this` parameter to these structures when creating a class and not when creating an interface, but there is no type-level enforcement that this happens.

This refactors the internals of the `class_sig` module to require a `this` type to be provided when creating a class to make this invariant less fragile.

Reviewed By: mvitousek

Differential Revision: D24504699

fbshipit-source-id: 9af84e8c91b745cafbeda21eb6798f60edd50994