Fix node/buffer libdef
Summary:
Module "buffer" in lib/node.js used to overwrite the definition of `Buffer`
from the toplevel. The local definition typed `Buffer` as `any`, and so every
reference to it as a type raised a `value-as-type` error. This diff assigns
the external type of `Buffer` to `Node$Buffer`, to avoid the name clash.

Reviewed By: mroch

Differential Revision: D25799625

fbshipit-source-id: 1c90fd2335748283a0110908be474f42cea2f94b