BETA: autoimports in autocomplete and code actions
Summary:
suggests exported symbols from across the codebase in autocomplete. if you select one, it automatically adds or updates the `import` statement.

also adds quick-fix code actions to add an import for undefined identifiers.

we attempt to match the user's style in terms of import formatting. if the existing imports are broken into sections, we add the new one to the correct section. if they're already sorted, we maintain the sort order. this reduces the chances that we conflict with eslint's [sort-imports](https://eslint.org/docs/rules/sort-imports) rule, without having to check its config.

NOTE: this is disabled by default so that it can be beta tested. set `autoimports=true` in flowconfig to enable per project.

also, since `flow autocomplete`'s legacy API doesn't support additional edits, which are necessary to insert the `import`, that API does not suggest autoimports by default even if they're enabled; pass `--imports` to include them.

Reviewed By: vrama628

Differential Revision: D25418703

fbshipit-source-id: f1277a5f512fe91ae60c39de65694340681c3bf9