fix autocomplete on opaque types
Summary:
autocomplete did not work on opaque types; it always returned `[]`. now it properly treats the type as not opaque in the file it was defined in, and as the super type if there is one in other files.

this diff adds support for opaque types to `Ty_normalizer.ExpandMembersConverter`. if the type is defined in a different file, we now recurse into its super type if it has one, or return an empty object. if it's defined in the same file, we recurse into its underlying type as if it was not opaque.

Reviewed By: vrama628, panagosg7

Differential Revision: D25925234 (https://github.com/facebook/flow/commit/0de4b2050eb2a8cc7e53b6e737ae4942bff19229)

fbshipit-source-id: f76c1792b7b13a40c2a60fb09997752f795839b9