raise EToplevelLibraryImport in statement.ml
Summary:
D25870407 (https://github.com/facebook/flow/commit/5148a475c0fe77e7a0d5c561ff0ee73070fd5fb1) banned toplevel import statement from libfiles. However, it still made
it possible for MergedT to appear when the respective types were optimized,
because the error was raised at the wrong spot.

The problem with the implementation in that diff is that the type assigned to such
bindings was still under-constrained: we would use some tvar that would await for a
lower-bound to be instantiated with, but that lower-bound would never arrive.
Later on, such tvars would be replaced with `MergedT` when optimized.

To avoid the creation of MergedT, we need to replace the under-constrained tvar
with some other type. This diff replaces it with `any`, as is typical in similar cases.

Reviewed By: jbrown215

Differential Revision: D25962397

fbshipit-source-id: ea9b538944712fa1fd7696daf7a0fc8863a56c80