replace uses of MergedT
Summary:
This diff replaces the construction of MergedT with `any`. There are a couple of
places where MergedT used to be created:

* Signature optimizer. This case is accompanied with a "missing-annotation" error.
This error shouldn't happen given that we input positions in exports need to be
annotated, but we're flagging the error for completeness.
* Type resolved for post-inference checks. Here we resort to `any` without an error.
Raising an error here would introduce a lot of noise. On the other hand, the cases
where `any` is added now, are unlikely to lead to any interesting check. These are
cases like untyped function parameters, etc, that (for now) we consider as typed as
 `empty`.

The next diff clears up MergedT from the codebase.

Reviewed By: jbrown215

Differential Revision: D25933829

fbshipit-source-id: 328c4930a9fe62d6b11e4b3e697165253a0fc90e