build Linux binaries on centos7
Summary:
the prebuilt Linux binaries were built on a debian8 VM. this was released in 2015, but EOL in June 2020. meanwhile, centos7 was released in 2014 and is supported until 2024.

Flow v0.143.0 includes some C++ code that, when built in debian8, requires a newer libc++ than centos7 has.

switching the build to centos7 restores portability.

Pull Request resolved: https://github.com/facebook/flow/pull/8576

Reviewed By: Hans-Halverson

Differential Revision: D26006003 (https://github.com/facebook/flow/commit/2c68cd0f33c067ea32218f1ee2a747bece2665db)

Pulled By: mroch

fbshipit-source-id: b0a23e4221bbf2b3d3e2f86f58fcfd3155e5f20b