[new-signatures] merge should produce ThisTypeT with is_this=false
Summary:
When merging class signatures the result is intended for external use. So, the
`is_this` portion of `ThisClassT` should be `false`.

Before in new-signatures mode, this was `true` which resulted in `this` not being
substituted to the class instance and instead remaining a `GenericT` outside the
scope of the class. In code like
```
// export.js
declare export class A4 { foo(): this; }
export default { A4 };

// import.js
import Export from './export';
declare var a4: Export.A4;
let _ = a4.foo();
```
this caused "escaped-generic" errors:
```
Error --------------------------------------------------------------------------------------------------- import.js:27:9

Cannot assign `a4.foo()` to `_` because `this` [1] contains `this` [2] which cannot escape from its class [3] (try
adding a type annotation to `_` [4]). [escaped-generic]

   import.js:27:9
   27| let _ = a4.foo();
               ^^^^^^^^

References:
   export.js:15:10
   15|   foo(): this;
                ^^^^ [1]
   import.js:26:17
   26| declare var a4: Export.A4;
                       ^^^^^^^^^ [2]
   export.js:14:22
   14| declare export class A4 {
                            ^^ [3]
   import.js:27:5
   27| let _ = a4.foo();
           ^ [4]

```

Reviewed By: mvitousek

Differential Revision: D26024217

fbshipit-source-id: c8d7af3593775e984c1f63fe506c43a3b2a1fef6