[parser] improve error recovery for declare opaque type with initializers
Summary:
`declare opaque type` doesn't allow initializers, as in `declare opaque type T : string = "foo";`

the error recovery was pretty garbage. since this is a common mistake, we now have a custom better error message, and parse the (invalid) initializer so that the next statement is not impacted.

if the next token is a semicolon or on a different line, then we predict it's a new statement and don't parse it as an initializer. this makes it recover better as you're typing in your IDE:

```
   declare opaque type Foo: number =
//                                 ^ error
   declare opaque type Bar: number;
// ^ successfully recovered

Reviewed By: Hans-Halverson

Differential Revision: D26150847

fbshipit-source-id: 21bd41a415bff640f098dee077c0811bc0e7e344