[parser] improve error recovery for declare opaque type with initializers
Summary:
`declare opaque type` doesn't allow initializers, as in `declare opaque type T : string = "foo";`

the error recovery was pretty garbage. since this is a common mistake, we now have a custom better error message, and parse the (invalid) initializer so that the next statement is not impacted.

if the next token is a semicolon or on a different line, then we predict it's a new statement and don't parse it as an initializer. this makes it recover better as you're typing in your IDE:

```
   declare opaque type Foo: number =
//                                 ^ error
   declare opaque type Bar: number;
// ^ successfully recovered

Reviewed By: Hans-Halverson

Differential Revision: D26150847 (https://github.com/facebook/flow/commit/377410cd77254f13c0a2f51af5972f15fb9e122b)

fbshipit-source-id: 41399745684a84cbff8395b229da849113f9b87a