[parser] improve JSX error handling
Summary:
Improves the parser's recovery when encountering partially-written JSX opening or closing tags, as you might experience while typing.

At a high level, the problem is that we currently use `Expect.token` too much. When it encounters an unexpected token, it both errors AND eats the bad token. This is important so that the parser continues to make progress, but only needs to happen once for each construct.

Now, we `Expect.token` for the `<`, but if anything else is wrong, we error and bail out of the element without consuming any more.

For example, take `<foo><bar</foo>`. Before, we hit the unexpected `<` and consume it before returning a best-effort `bar` child, which prevents us from properly parsing the closing `foo` tag, leading to runaway errors.

Now, we error on but don't consume the `<`, returning the same best-effort `bar` child. But now it recovers properly and sees the entire `</foo>` closing tag.

Reviewed By: gkz

Differential Revision: D26051008 (https://github.com/facebook/flow/commit/b45fa1cf5403d24fdf5f78c6f67763f6b31040f0)

fbshipit-source-id: 348125855b80f9e6f5741008a7d9d8011a03e2ac