[parser] rename Expect.maybe to Eat.maybe
Summary:
`Expect.token` errors if the next token isn't what was expected. `Eat.token` eats the next token regardless.

`Expect.maybe` eats the next token if it was expected, but does not error if it wasn't. This is more similar to `Eat` than `Expect`. if `if` wasn't a reserved word, think of it as `Eat.if env T_FOO`.

so I renamed `Expect.maybe` to `Eat.token`, and intend to add a new `Expect.maybe` that errors like `Expect.token`, but doesn't consume.

Reviewed By: gkz

Differential Revision: D26151838 (https://github.com/facebook/flow/commit/fd419c25f457149de7835ca7d2bc994d3713debd)

fbshipit-source-id: 22975b91a84c1f49a3fdb373df392cdab73d953c