ThisClassT lower bound in implicit instantiations
Summary:
extends our implicit instantiation to classes. It's much simpler to calculate the variance of type parameters here because we can take the variance sigil (or lack thereof) from the polymorphic declaration. After that, we can just send it through the same process we do for functions by calling the constructor and picking bounds.

Note on interfaces: Interfaces with type parameters are only types, they can't be referenced directly as a value. For example:
```
interface A<T> {
  constructor(...): this
}
new A(); // Error, A is not a value
```

So the only way to call the constructor on this interface is to have a value that is annotated with the interface, and any annotation would have to also include a type argument for the interfaces type parameters. Any polymorphic functions bound to the interface would go through the normal function instantiation.

As such, I did not include tests for interfaces.

Reviewed By: panagosg7

Differential Revision: D25974516

fbshipit-source-id: 6c435a415b6905d52063dacc6a72c1ac8aa4ccda