Pin and print types in implicit instantiations
Summary:
With the polarity information from the visitor, we finish implementing Pierce's algorithm for implicit instantiations in this diff:

If the type param is in a covariant position in the return, use the lower bounds we get from the call

If the type param is in a contravariant position in the return, then use the upper bounds we get from the call

If the type param is in an invariant position in the return, then unify the lower + upper bounds and use that. In this diff, we don't actually perform the unification and just take the lower bounds so that we can avoid polluting the error output (which is where we get all the information about what Flow is inferring).

This does not take into account any complications presented by interdependent bounds.

Reviewed By: panagosg7

Differential Revision: D24866678

fbshipit-source-id: f365bc3ecf6fb72a85131553529dde6ad21638ba